fix(review): satisfy classification review gates
