+++ pkg/services/factory_runtime/internal/services/orchestration/subsystems/subsystem_transitioner_test.go
@@ -413,35 +413,6 @@ func TestResolveWorkResult_RuntimeConfigStopWordsFailMissingMarker(t *testing.T)
 	}
 }
 
-func TestResolveWorkResult_DecisionEnvelopePreservesExplicitRejectionOverStopWords(t *testing.T) {
-	transition := &petri.Transition{
-		ID:   "transition-id",
-		Name: "review-station",
-	}
-	result := &workerexecution.WorkResult{
-		DispatchID:   "dispatch-1",
-		TransitionID: "transition-id",
-		Outcome:      workerexecution.OutcomeRejected,
-		Feedback:     "add the missing release date",
-	}
-
-	resolved := resolveWorkResult(transition, result, runtimefixtures.RuntimeWorkstationLookupFixture{
-		Workstations: map[string]*interfaces.FactoryWorkstationConfig{
-			"review-station": {
-				OutcomeFormat: interfaces.WorkstationOutcomeFormatDecisionEnvelope,
-				StopWords:     []string{"DONE"},
-			},
-		},
-	})
-
-	if resolved.outcome != workerexecution.OutcomeRejected {
-		t.Fatalf("resolved outcome = %s, want REJECTED", resolved.outcome)
-	}
-	if resolved.feedback != result.Feedback {
-		t.Fatalf("resolved feedback = %q, want %q", resolved.feedback, result.Feedback)
-	}
-}
-
 func TestResolveWorkResult_MissingRuntimeConfigPreservesOriginalOutcome(t *testing.T) {
 	transition := &petri.Transition{
 		ID: "runtime-station-id",
